// frontend/generated/api.ts
export interface PlayerDTO {
    'id'?: number;
    'tag'?: string;
    'name'?: string;
    'trophies'?: number;
    'highestTrophies'?: number;
    'expLevel'?: number;
    'expPoints'?: number;
    'totalPrestigeLevel'?: number;
    'soloVictories'?: number;
    'duoVictories'?: number;
    'threeVsThreeVictories'?: number;
    'bestRoboRumbleTime'?: number;
    'bestTimeAsBigBrawler'?: number;
    'isQualifiedFromChampionshipChallenge'?: boolean;
    'clubTag'?: string;
    'clubName'?: string;
    'createdAt'?: string;
}